refactor: move e2e tests to the folder
